initial refactor for work on llm evaluation
